requirements.txt
asyncpg~=0.21.0
discord.py[voice]==1.5.0
lavalink~=3.1.2
psutil==5.7.2
python-dotenv~=0.14.0
uvloop==0.14.0